feat(chore): Remaniement, regroupement des fonctions de log
